+++ BillSplitter/iOSProjectTemplate/Constants.h
@@ -9,7 +9,7 @@
 #import <Foundation/Foundation.h>
 
 // UI
-extern float const UI_SIZE_MIN_TOUCH;
+extern CGFloat const UI_SIZE_MIN_TOUCH;
 
 // Cached Data
 extern NSString* const CACHE_KEY_USER_SETTINGS;
@@ -23,13 +23,13 @@ extern NSString* const FONT_NAME_HEADERS;
 extern NSString* const FONT_NAME_COPY;
 extern NSString* const FONT_NAME_TEXTFIELD;
 extern NSString* const FONT_NAME_TAGLINE;
-extern float const FONT_SIZE_TAGLINE;
-extern float const FONT_SIZE_COPY;
-extern float const FONT_SIZE_SMALL_LABEL;
-extern float const FONT_SIZE_HEADCOUNT;
-extern float const FONT_SIZE_PRICE;
-extern float const FONT_SIZE_BRANDING;
-extern float const FONT_SIZE_TITLE;
+extern CGFloat const FONT_SIZE_TAGLINE;
+extern CGFloat const FONT_SIZE_COPY;
+extern CGFloat const FONT_SIZE_SMALL_LABEL;
+extern CGFloat const FONT_SIZE_HEADCOUNT;
+extern CGFloat const FONT_SIZE_PRICE;
+extern CGFloat const FONT_SIZE_BRANDING;
+extern CGFloat const FONT_SIZE_TITLE;
 
 // Time
 extern int const TIME_ONE_MINUTE;
@@ -51,9 +51,9 @@ extern int const COLOR_HEX_ACCENT;
 extern int const COLOR_HEX_LIGHT_ACCENT;
 
 // Animations
-extern float const ANIMATION_DURATION_FASTEST;
-extern float const ANIMATION_DURATION_FAST;
-extern float const ANIMATION_DURATION_MED;
-extern float const ANIMATION_DURATION_SLOW;
-extern float const ANIMATION_DURATION_SLOWEST;
+extern CGFloat const ANIMATION_DURATION_FASTEST;
+extern CGFloat const ANIMATION_DURATION_FAST;
+extern CGFloat const ANIMATION_DURATION_MED;
+extern CGFloat const ANIMATION_DURATION_SLOW;
+extern CGFloat const ANIMATION_DURATION_SLOWEST;
